requirements.txt
python-binance
python-dotenv
requests
apscheduler